requirements.txt
PySide6
pandas
scipy
liac-arff
